requirements.txt
Pillow
numpy
opencv-python
opencv-contrib-python
